@@ -60,15 +60,13 @@ public final class URLDependencyDownloader implements DependencyDownloader {
     @Override
     public File download(final Dependency dependency) throws IOException {
         final File expectedOutputFile = outputWriterProducer.getStrategy().selectFileFor(dependency);
-        if (expectedOutputFile.exists()
-            &&  expectedOutputFile.length() == BOM_BYTES.length
-            &&  Arrays.equals(Files.readAllBytes(expectedOutputFile.toPath()), BOM_BYTES)
-        ) {
-            return null;
-        }
-        if (verifier.verify(expectedOutputFile, dependency)) {
-            return expectedOutputFile;
-        }
+
+        if (expectedOutputFile.exists() &&
+            expectedOutputFile.length() == BOM_BYTES.length &&
+            Arrays.equals(Files.readAllBytes(expectedOutputFile.toPath()), BOM_BYTES)
+        ) return null;
+
+        if (verifier.verify(expectedOutputFile, dependency)) return expectedOutputFile;
 
         final ResolutionResult result = dependencyResolver.resolve(dependency)
                 .orElseThrow(() -> new UnresolvedDependencyException(dependency));
@@ -86,21 +84,24 @@ public final class URLDependencyDownloader implements DependencyDownloader {
             checksumFile.delete();
         }
 
-        LOGGER.log("Downloading {0}...", dependency.getArtifactId());
+        LOGGER.log("Downloading %s...", dependency.artifactId());
 
         final URL url = result.getDependencyURL();
-        LOGGER.debug("Connecting to {0}", url);
+        LOGGER.debug("Connecting to %s", url);
+
         final URLConnection connection = Connections.createDownloadConnection(url);
         final InputStream inputStream = connection.getInputStream();
-        LOGGER.debug("Connection successful! Downloading {0}" ,dependency.getArtifactId() + "...");
+        LOGGER.debug("Connection successful! Downloading %s" ,dependency.artifactId() + "...");
+
         final OutputWriter outputWriter = outputWriterProducer.create(dependency);
-        LOGGER.debug("{0}.Size = {1}", dependency.getArtifactId(), connection.getContentLength());
+        LOGGER.debug("%s.Size = %s", dependency.artifactId(), connection.getContentLength());
+
         final File downloadResult = outputWriter.writeFrom(inputStream, connection.getContentLength());
         Connections.tryDisconnect(connection);
         verifier.verify(downloadResult, dependency);
-        LOGGER.debug("Artifact {0} downloaded successfully!", dependency.getArtifactId());
+        LOGGER.debug("Artifact %s downloaded successfully!", dependency.artifactId());
 
-        LOGGER.log("Downloaded {0} successfully!", dependency.getArtifactId());
+        LOGGER.log("Downloaded %s successfully!", dependency.artifactId());
         return downloadResult;
     }
 }
