@@ -74,8 +74,8 @@ public final class InjectionHelperFactory {
 
     public InjectionHelper create(final DependencyData data, final Map<String, ResolutionResult> preResolvedResults) throws IOException, NoSuchAlgorithmException, URISyntaxException {
         final Collection<Repository> repositories = mirrorSelector
-                .select(data.getRepositories(), data.getMirrors());
-        final Relocator relocator = relocatorFactory.create(data.getRelocations());
+                .select(data.repositories(), data.mirrors());
+        final Relocator relocator = relocatorFactory.create(data.relocations());
         final RelocationHelper relocationHelper = relocationHelperFactory.create(relocator);
         final FilePathStrategy filePathStrategy = FilePathStrategy.createDefault(downloadDirectoryPath.toFile());
         final OutputWriterFactory outputWriterFactory = new DependencyOutputWriterFactory(filePathStrategy);
