@@ -28,13 +28,19 @@ import java.util.Collection;
 import java.util.Objects;
 
 public final class Dependency {
-    private final String groupId;
+    private String groupId;
     private final String artifactId;
     private final String version;
     private final String snapshotId;
     private final Collection<Dependency> transitive;
 
-    public Dependency(String groupId, String artifactId, String version, String snapshotId, Collection<Dependency> transitive) {
+    public Dependency(
+        String groupId,
+        String artifactId,
+        String version,
+        String snapshotId,
+        Collection<Dependency> transitive
+    ) {
         this.groupId = groupId;
         this.artifactId = artifactId;
         this.version = version;
@@ -42,31 +48,32 @@ public final class Dependency {
         this.transitive = transitive;
     }
 
-    public String getGroupId() {
+
+    public String groupId() {
         return groupId;
     }
 
-    public String getArtifactId() {
+    public String artifactId() {
         return artifactId;
     }
 
-    public String getVersion() {
+    public String version() {
         return version;
     }
 
-    public String getSnapshotId() {
+    public String snapshotId() {
         return snapshotId;
     }
 
-    public Collection<Dependency> getTransitive() {
+    public Collection<Dependency> transitive() {
         return transitive;
     }
 
     @Override
     public String toString() {
-        final String snapshotId = getSnapshotId();
-        final String suffix = (snapshotId != null && snapshotId.length() > 0) ? (":" + snapshotId): "";
-        return getGroupId() + ":" + getArtifactId() + ":" + getVersion() + suffix;
+        final String snapshotId = snapshotId();
+        final String suffix = (snapshotId != null && snapshotId.length() > 0) ? (":" + snapshotId) : "";
+        return groupId() + ":" + artifactId() + ":" + version() + suffix;
     }
 
     @Override
@@ -75,13 +82,12 @@ public final class Dependency {
         if (o == null || getClass() != o.getClass()) return false;
         Dependency that = (Dependency) o;
         return groupId.equals(that.groupId) &&
-                artifactId.equals(that.artifactId) &&
-                version.equals(that.version);
+            artifactId.equals(that.artifactId) &&
+            version.equals(that.version);
     }
 
     @Override
     public int hashCode() {
         return Objects.hash(groupId, artifactId, version);
     }
-
 }
