@@ -27,30 +27,17 @@ package io.github.slimjar.resolver.data;
 import java.net.URL;
 import java.util.Objects;
 
-public final class Mirror {
-
-    private final URL mirroring;
-    private final URL original;
-
-    public Mirror(URL mirroring, URL original) {
-        this.mirroring = mirroring;
-        this.original = original;
-    }
-
-    public URL getMirroring() {
-        return mirroring;
-    }
-
-    public URL getOriginal() {
-        return original;
-    }
+public record Mirror(
+    URL mirroring,
+    URL original
+) {
 
     @Override
     public String toString() {
         return "Mirror{" +
-                "mirror=" + mirroring +
-                ", original=" + original +
-                '}';
+            "mirror=" + mirroring +
+            ", original=" + original +
+            '}';
     }
 
     @Override
@@ -59,11 +46,6 @@ public final class Mirror {
         if (o == null || getClass() != o.getClass()) return false;
         final Mirror mirror1 = (Mirror) o;
         return Objects.equals(mirroring, mirror1.mirroring) && Objects.equals(original, mirror1.original);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(mirroring, original);
     }
 
 }
