@@ -40,7 +40,7 @@ public final class MediatingPathResolutionStrategy implements PathResolutionStra
 
     @Override
     public Collection<String> pathTo(final Repository repository, final Dependency dependency) {
-        if (dependency.getSnapshotId() != null) {
+        if (dependency.snapshotId() != null) {
             return snapshotStrategy.pathTo(repository, dependency);
         }
         return releaseStrategy.pathTo(repository, dependency);
