@@ -28,17 +28,25 @@ import io.github.slimjar.resolver.data.Dependency;
 import io.github.slimjar.resolver.data.Repository;
 import io.github.slimjar.resolver.enquirer.RepositoryEnquirerFactory;
 import io.github.slimjar.resolver.pinger.URLPinger;
-import junit.framework.TestCase;
-import org.junit.Assert;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Optional;
+import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.mockito.Mockito;
 
-public class DependencyResolverTest extends TestCase {
+public class DependencyResolverTest {
 
+    @AfterEach
+    public void tearDown() {
+        Mockito.clearAllCaches();
+    }
+
+    @Test
     public void testCachingDependencyResolverResolution() throws MalformedURLException {
         final RepositoryEnquirerFactory repositoryEnquirerFactory = new DummyRepositoryEnquirerFactory();
         final Collection<Repository> repositories = Collections.singleton(new Repository(new URL("https://repo.tld/")));
@@ -55,10 +63,11 @@ public class DependencyResolverTest extends TestCase {
         }, repositories, repositoryEnquirerFactory, Collections.emptyMap());
         final Dependency testDependency = new Dependency("a.b.c", "d", "1.0", null, Collections.emptyList());
         final Optional<ResolutionResult> url = dependencyResolver.resolve(testDependency);
-        final String groupPath = testDependency.getGroupId().replace('.', '/');
-        Assert.assertEquals(url.get().getDependencyURL().toString(), String.format("https://repo.tld/%s/%s/%s/%2$s-%3$s.jar", groupPath, testDependency.getArtifactId(), testDependency.getVersion()));
+        final String groupPath = testDependency.groupId().replace('.', '/');
+        Assertions.assertEquals(url.get().getDependencyURL().toString(), String.format("https://repo.tld/%s/%s/%s/%2$s-%3$s.jar", groupPath, testDependency.artifactId(), testDependency.version()));
     }
 
+    @Test
     public void testCachingDependencyResolverCaching() throws MalformedURLException {
         final RepositoryEnquirerFactory repositoryEnquirerFactory = new DummyRepositoryEnquirerFactory();
         final Collection<Repository> repositories = Collections.singleton(new Repository(new URL("https://repo.tld/")));
@@ -77,7 +86,7 @@ public class DependencyResolverTest extends TestCase {
         final URL url1 = dependencyResolver.resolve(testDependency).get().getDependencyURL();
         final URL url2 = dependencyResolver.resolve(testDependency).get().getDependencyURL();
 
-        Assert.assertEquals("Impure resolution: Instance", url1, url2);
+        Assertions.assertEquals(url1, url2, "Impure resolution: Instance");
     }
 
 }
