@@ -26,24 +26,27 @@ package io.github.slimjar.resolver.mirrors;
 
 import io.github.slimjar.resolver.data.Mirror;
 import io.github.slimjar.resolver.data.Repository;
-import junit.framework.TestCase;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Collection;
 import java.util.Collections;
+import org.junit.jupiter.api.Assertions;
+import org.junit.Test;
 
-public class MirrorSelectorTest extends TestCase {
+public class MirrorSelectorTest {
+    @Test
     public void testSelectorForceMirrorCentral() throws MalformedURLException {
         final Repository central = new Repository(new URL(SimpleMirrorSelector.CENTRAL_URL));
         final Repository centralMirror = new Repository(new URL(SimpleMirrorSelector.DEFAULT_CENTRAL_MIRROR_URL));
         final Collection<Repository> original = Collections.singleton(central);
         final MirrorSelector mirrorSelector = new SimpleMirrorSelector(Collections.singleton(centralMirror));
         final Collection<Repository> selected = mirrorSelector.select(original, Collections.emptyList());
-        assertFalse("Selection should remove central", selected.contains(central));
-        assertTrue("Selection should contain central mirror", selected.contains(new Repository(new URL(SimpleMirrorSelector.DEFAULT_CENTRAL_MIRROR_URL))));
+        Assertions.assertFalse(selected.contains(central), "Selection should remove central");
+        Assertions.assertTrue(selected.contains(new Repository(new URL(SimpleMirrorSelector.DEFAULT_CENTRAL_MIRROR_URL))), "Selection should contain central mirror");
     }
 
+    @Test
     public void testSelectorForceAltMirrorCentral() throws MalformedURLException {
         final Repository central = new Repository(new URL(SimpleMirrorSelector.CENTRAL_URL));
         final Repository centralMirror = new Repository(new URL(SimpleMirrorSelector.DEFAULT_CENTRAL_MIRROR_URL));
@@ -52,23 +55,24 @@ public class MirrorSelectorTest extends TestCase {
         final MirrorSelector mirrorSelector = new SimpleMirrorSelector(Collections.singleton(centralMirror));
 
         final Collection<Repository> selected = mirrorSelector.select(original, Collections.emptyList());
-        assertFalse("Selection should remove central", selected.contains(central));
-        assertTrue("Selection should contain central mirror", selected.contains(new Repository(new URL(SimpleMirrorSelector.DEFAULT_CENTRAL_MIRROR_URL))));
+        Assertions.assertFalse(selected.contains(central), "Selection should remove central");
+        Assertions.assertTrue(selected.contains(new Repository(new URL(SimpleMirrorSelector.DEFAULT_CENTRAL_MIRROR_URL))), "Selection should contain central mirror");
     }
 
+    @Test
     public void testSelectorReplace() throws MalformedURLException {
         final Repository central = new Repository(new URL(SimpleMirrorSelector.CENTRAL_URL));
         final Collection<Repository> centralMirrors = Collections.singleton(central);
         final Repository originalRepo = new Repository(new URL("https://a.b.c"));
         final Repository mirroredRepo = new Repository(new URL("https://d.e.f"));
-        final Mirror mirror = new Mirror(mirroredRepo.getUrl(), originalRepo.getUrl());
+        final Mirror mirror = new Mirror(mirroredRepo.url(), originalRepo.url());
         final Collection<Repository> original = Collections.singleton(originalRepo);
         final Collection<Mirror> mirrors = Collections.singleton(mirror);
         final MirrorSelector mirrorSelector = new SimpleMirrorSelector(centralMirrors);
 
 
         final Collection<Repository> selected = mirrorSelector.select(original, mirrors);
-        assertFalse("Selection should remove original", selected.contains(originalRepo));
-        assertTrue("Selection should contain mirror", selected.contains(mirroredRepo));
+        Assertions.assertFalse(selected.contains(originalRepo), "Selection should remove original");
+        Assertions.assertTrue(selected.contains(mirroredRepo), "Selection should contain mirror");
     }
 }
